Refactored the database to be used with bungee, spigot and sponge
