+++ multichat/src/main/java/xyz/olivermartin/multichat/database/DatabaseMode.java
@@ -1,0 +1,8 @@
+package xyz.olivermartin.multichat.database;
+
+public enum DatabaseMode {
+
+	SQLite,
+	MySQL
+	
+}
