+++ multichat/src/main/java/xyz/olivermartin/multichat/database/GenericDatabase.java
@@ -1,4 +1,4 @@
-package xyz.olivermartin.multichat.spigotbridge.database;
+package xyz.olivermartin.multichat.database;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
