+++ multichat/src/main/java/xyz/olivermartin/multichat/database/MySQLDatabase.java
@@ -1,4 +1,4 @@
-package xyz.olivermartin.multichat.spigotbridge.database;
+package xyz.olivermartin.multichat.database;
 
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
