+++ multichat/src/main/java/xyz/olivermartin/multichat/database/SQLiteDatabase.java
@@ -1,4 +1,4 @@
-package xyz.olivermartin.multichat.spigotbridge.database;
+package xyz.olivermartin.multichat.database;
 
 import java.io.File;
 import java.sql.Connection;
