--- multichat/src/main/java/xyz/olivermartin/multichat/spigotbridge/database/DatabaseMode.java
@@ -1,8 +1,0 @@
-package xyz.olivermartin.multichat.spigotbridge.database;
-
-public enum DatabaseMode {
-
-	SQLite,
-	MySQL
-	
-}
